Optinal usage of DTK Action Agent sync, based on server's conf

+++ mcollective_additions/plugins/v2.2/agent/dev_manager.rb
@@ -34,26 +34,28 @@ module MCollective
           dtk_aa_branch = request[:action_agent_branch]
 
           # DTK Action Agent Sync
-          begin
-            Log.instance.info("Started DTK Action Agent sync, temp dir '#{DTK_AA_TEMP_DIR}'")
-            output = `git clone #{dtk_aa_url} -b #{dtk_aa_branch} #{DTK_AA_TEMP_DIR}`
-            result = $?
-
-            if result.success?
-              Log.instance.info("Cloned latest code from branch '#{dtk_aa_branch}' for DTK Action Agent.")
-
-              output = `cd #{DTK_AA_TEMP_DIR} && /opt/puppet-omnibus/embedded/bin/gem build #{DTK_AA_TEMP_DIR}/dtk-action-agent.gemspec && /opt/puppet-omnibus/embedded/bin/gem install #{DTK_AA_TEMP_DIR}/dtk-action-agent-*.gem --no-ri --no-rdoc`
+          if dtk_aa_url && dtk_aa_branch
+            begin
+              Log.instance.info("Started DTK Action Agent sync, temp dir '#{DTK_AA_TEMP_DIR}'")
+              output = `git clone #{dtk_aa_url} -b #{dtk_aa_branch} #{DTK_AA_TEMP_DIR}`
               result = $?
+
               if result.success?
-                Log.instance.info("DTK Action Agent has been successfully update from branch '#{dtk_aa_branch}'")
+                Log.instance.info("Cloned latest code from branch '#{dtk_aa_branch}' for DTK Action Agent.")
+
+                output = `cd #{DTK_AA_TEMP_DIR} && /opt/puppet-omnibus/embedded/bin/gem build #{DTK_AA_TEMP_DIR}/dtk-action-agent.gemspec && /opt/puppet-omnibus/embedded/bin/gem install #{DTK_AA_TEMP_DIR}/dtk-action-agent-*.gem --no-ri --no-rdoc`
+                result = $?
+                if result.success?
+                  Log.instance.info("DTK Action Agent has been successfully update from branch '#{dtk_aa_branch}'")
+                else
+                  Log.instance.error("DTK Action Agent could not be updated, reason: #{output}")
+                end
               else
-                Log.instance.error("DTK Action Agent could not be updated, reason: #{output}")
+                Log.instance.error("Not able to clone latest code from '#{dtk_aa_url} branch #{dtk_aa_branch}', aborting DTK Action Agent Sync. #{output}")
               end
-            else
-              Log.instance.error("Not able to clone latest code from '#{dtk_aa_url} branch #{dtk_aa_branch}', aborting DTK Action Agent Sync. #{output}")
+            ensure
+              FileUtils.rm_rf DTK_AA_TEMP_DIR
             end
-          ensure
-            FileUtils.rm_rf DTK_AA_TEMP_DIR
           end
 
           t1 = Thread.new do
